Adding --delete_fields flag as option

+++ migrate_fields_config.py
@@ -194,6 +194,7 @@ if __name__ == '__main__':
     parser.add_argument('--v2_org_id', required=True)
     parser.add_argument('--v2_source_id', required=True)
     parser.add_argument('--v2_access_token', required=True)
+    parser.add_argument('--delete_fields', action='store_true')
     opts = parser.parse_args()
 
     # args
@@ -204,6 +205,7 @@ if __name__ == '__main__':
     v2_org_id = opts.v2_org_id
     v2_source_id = opts.v2_source_id
     v2_access_token = opts.v2_access_token
+    delete_fields = opts.delete_fields
 
     v1_client = CloudV1(env, v1_org_id, v1_access_token)
     v1_sources = v1_client.get_sources()
@@ -232,7 +234,8 @@ if __name__ == '__main__':
         print('No fields to update.')
     print('Migration completed.')    
     #Deleting unused fields
-    unusedfields = get_unused_fields(v2_client.get_fields_with_mappings())
-    print('Deleting fields ' + unusedfields)
-    v2_client.delete_fields(unusedfields)
-    print('Field deletion completed.')
+    if delete_fields:
+        unusedfields = get_unused_fields(v2_client.get_fields_with_mappings())
+        print('Deleting fields ' + unusedfields)
+        v2_client.delete_fields(unusedfields)
+        print('Field deletion completed.')
